long integers for handling size
git-svn-id: file:///home/roberto/Git/svn@1223 1a721e8d-d38c-4e91-ab98-6a0be4d754af

+++ trunk/Xmipp/Lib/XmippData/Src/MultidimBasic.hh
@@ -412,7 +412,7 @@ public:
    void core_init() {__m=NULL; __dim=0;}
 
    /// Allocate vector memory
-   void core_allocate(int _dim) {
+   void core_allocate(long int _dim) {
       if (_dim<=0) {clear(); return;}
       __m = new T [_dim];
       if (__m != NULL) __dim = _dim;
